+++ codex/serializers/choices.py
@@ -48,7 +48,7 @@ _CHOICES_FN_RE = {
     _CHOICES_MODULE_NAME: create_choices_fn_regexes(_CHOICES_MODULE_NAME),
     _ADMIN_CHOICES_MODULE_NAME: create_choices_fn_regexes(_ADMIN_CHOICES_MODULE_NAME),
 }
-_SINGLE_VALUE_KEYS = frozenset({"q", "route"})
+_SINGLE_VALUE_KEYS = frozenset({"q", "route", "searchResultsLimit"})
 
 # Exports
 CHOICES = {}