+++ frontend/src/components/browser/drawer/search-help.vue
@@ -144,6 +144,14 @@
           clear the filter selector or the search field if you prefer to apply
           only one of them.
         </p>
+        <h2>Faster Browsing and Searching</h2>
+        <p>
+          Browsing large collections using the group tree and filter dropdown is
+          relatively fast. Searching with the search bar will always be slower.
+          To aid search speed there is a setting for limiting search results
+          which can help with large collections. To see all results, choose
+          Search Result Limit: "Unlimited".
+        </p>
       </div>
       <CloseButton
         class="closeButton"